Move mock server handlers to separate file
Use `index.ts` as an actual index that re-exports other files in the
directory, like in other places in our codebase.
